Completely redesign api
The API has been completely redesigned and should now offer all the
stats in a format similar to their datatype. Take a look at
ResultSetConveter.java for more information. It also respects your
disabled columns setting in the config, so if you disable address and
player corrds, the api won't send that data.